Copy notebooks, sql magic, and prettytable script into Dockerfile
This moves the getting started ipython notebook and the sql
magic/prettytable script into the Dockerfile. This is so that
it's included when the image is published as `tabulario/spark-iceberg`.

The `docker-compose.yml` file was also updated to use the `tabulario/spark-iceberg`
image instead of building from the directories in this repo. Togther, this will let
someone run this setup with only a `docker-compose.yml` file (no need to clone this
repo).